Change precision to 64-bit
This was causing errors because Polylabel assumes 64-bit

+++ src/main.rs
@@ -64,7 +64,7 @@ fn open_and_parse(p: &str) -> Result<GeoJson, PolylabelError> {
 }
 
 /// Process top-level `GeoJSON` items
-fn process_geojson(gj: &mut GeoJson, tolerance: &f32, ctr: &AtomicIsize) {
+fn process_geojson(gj: &mut GeoJson, tolerance: &f64, ctr: &AtomicIsize) {
     match *gj {
         GeoJson::FeatureCollection(ref mut collection) => collection.features
             .par_iter_mut()
@@ -81,7 +81,7 @@ fn process_geojson(gj: &mut GeoJson, tolerance: &f32, ctr: &AtomicIsize) {
 }
 
 /// Process `GeoJSON` geometries
-fn label_geometry(geom: &mut Geometry, tolerance: &f32, ctr: &AtomicIsize) {
+fn label_geometry(geom: &mut Geometry, tolerance: &f64, ctr: &AtomicIsize) {
     match geom.value {
         Value::Polygon(_) | Value::MultiPolygon(_) => label_value(Some(geom), tolerance, ctr),
         Value::GeometryCollection(ref mut collection) => {
@@ -101,18 +101,18 @@ fn label_geometry(geom: &mut Geometry, tolerance: &f32, ctr: &AtomicIsize) {
 }
 
 /// Generate a label position for a (Multi)Polygon Value
-fn label_value(geom: Option<&mut Geometry>, tolerance: &f32, ctr: &AtomicIsize) {
+fn label_value(geom: Option<&mut Geometry>, tolerance: &f64, ctr: &AtomicIsize) {
     if let Some(gmt) = geom {
         // construct a fake empty Polygon – this doesn't allocate
         // TODO if Geo geometry validation lands, this will fail
-        let v1: Vec<Point<f32>> = Vec::new();
+        let v1: Vec<Point<f64>> = Vec::new();
         let ls2 = Vec::new();
-        let fake_polygon: Polygon<f32> = Polygon::new(LineString::from(v1), ls2);
+        let fake_polygon: Polygon<f64> = Polygon::new(LineString::from(v1), ls2);
         // convert it into a Value, and swap it for our actual (Multi)Polygon
         gmt.value = match gmt.value {
             Value::Polygon(_) => {
                 let intermediate = replace(&mut gmt.value, Value::from(&fake_polygon));
-                let geo_type: Polygon<f32> = intermediate
+                let geo_type: Polygon<f64> = intermediate
                     .try_into()
                     .expect("Failed to convert a Polygon");
                 // bump the Polygon counter
@@ -122,7 +122,7 @@ fn label_value(geom: Option<&mut Geometry>, tolerance: &f32, ctr: &AtomicIsize) 
             }
             Value::MultiPolygon(_) => {
                 let intermediate = replace(&mut gmt.value, Value::from(&fake_polygon));
-                let geo_type: MultiPolygon<f32> = intermediate
+                let geo_type: MultiPolygon<f64> = intermediate
                     .try_into()
                     .expect("Failed to convert a MultiPolygon");
                 // we allocate here – can we avoid it? idk
@@ -194,7 +194,7 @@ fn main() {
                 .required(true))
        .get_matches();
 
-    let tolerance = value_t!(command_params.value_of("TOLERANCE"), f32).unwrap_or(0.001);
+    let tolerance = value_t!(command_params.value_of("TOLERANCE"), f64).unwrap_or(0.001);
     let poly = value_t!(command_params.value_of("GEOJSON"), String).unwrap();
     let pprint = command_params.is_present("pretty");
     let statsonly = command_params.is_present("statsonly");
